datapath/loader: Unparallel TestSetupIPIPDevices
The `TestSetupIPIPDevices` test calls `setupIPIPDevices` which creates
IPIP{4,6} devices. This implicitly causes kernel modules for these
devices to be loaded. When that happens default behavior is to create
"fallback" devices, not only in the main network namespace, but also
in any network namespace that is created.

This behavior can be changed with the
`net.core.fb_tunnels_only_for_init_net` sysctl. In normal Cilium
operation we set this sysctl to `2` which disabled this behavior.
However, we have legacy/fallback code that renames these fallback
devices and to test that rename code we run this test with a sysctl
value of `0` which enables the fallback devices to be created in
all network namespaces.

This is problematic since it makes asserting on the state of the
device table in unrelated tests difficult. Depending on if this test
is executed in parallel with such tests or not. By default, `go test`
runs tests in separate packages in parallel.

Others have encountered similar issues in the past so we already have
the `//go:build unparallel` build tag. This commit moves the
`TestSetupIPIPDevices` test to a new file `netlink_unparallel_test.go`
and adds the `//go:build unparallel` build tag to it. This ensures that
this test is only ever run serialized.

I also added cleanup code to change the sysctl value to `0` at the
start of the test and to `2` at the end of the test. This ensures
we always have the expected state during and after the test.
