+++ pkg/datapath/loader/netlink_test.go
@@ -450,68 +450,3 @@ func TestAddHostDeviceAddr(t *testing.T) {
 		return nil
 	})
 }
-
-func TestSetupIPIPDevices(t *testing.T) {
-	testutils.PrivilegedTest(t)
-	logger := hivetest.Logger(t)
-
-	sysctl := sysctl.NewDirectSysctl(afero.NewOsFs(), "/proc")
-
-	ns := netns.NewNetNS(t)
-
-	ns.Do(func() error {
-		err := setupIPIPDevices(logger, sysctl, true, true, 1500)
-		require.NoError(t, err)
-
-		dev4, err := netlink.LinkByName(defaults.IPIPv4Device)
-		require.NoError(t, err)
-		require.Equal(t, 1480, dev4.Attrs().MTU)
-
-		dev6, err := netlink.LinkByName(defaults.IPIPv6Device)
-		require.NoError(t, err)
-		require.Equal(t, 1452, dev6.Attrs().MTU)
-
-		_, err = netlink.LinkByName("cilium_tunl")
-		require.NoError(t, err)
-
-		_, err = netlink.LinkByName("cilium_ip6tnl")
-		require.NoError(t, err)
-
-		_, err = netlink.LinkByName("tunl0")
-		require.Error(t, err)
-
-		_, err = netlink.LinkByName("ip6tnl0")
-		require.Error(t, err)
-
-		err = setupIPIPDevices(logger, sysctl, false, false, 1500)
-		require.NoError(t, err)
-
-		_, err = netlink.LinkByName(defaults.IPIPv4Device)
-		require.Error(t, err)
-
-		_, err = netlink.LinkByName(defaults.IPIPv6Device)
-		require.Error(t, err)
-
-		err = setupIPIPDevices(logger, sysctl, true, true, 1480)
-		require.NoError(t, err)
-
-		dev4, err = netlink.LinkByName(defaults.IPIPv4Device)
-		require.NoError(t, err)
-		require.Equal(t, 1460, dev4.Attrs().MTU)
-
-		dev6, err = netlink.LinkByName(defaults.IPIPv6Device)
-		require.NoError(t, err)
-		require.Equal(t, 1432, dev6.Attrs().MTU)
-
-		err = setupIPIPDevices(logger, sysctl, false, false, 1480)
-		require.NoError(t, err)
-
-		_, err = netlink.LinkByName(defaults.IPIPv4Device)
-		require.Error(t, err)
-
-		_, err = netlink.LinkByName(defaults.IPIPv6Device)
-		require.Error(t, err)
-
-		return nil
-	})
-}
